+++ src/cpp/RollingMedian.h
@@ -58,6 +58,82 @@ namespace OrderedStructs {
         };
 
 /**
+ * Vector based rolling median.
+ * This handles both even and odd window lengths.
+ *
+ * The length of the result is data.size() - win_length
+ *
+ * @tparam T Data type.
+ * @param data Data vector.
+ * @param win_length Window length.
+ * @param result Result vector.
+ * @return ROLLING_MEDIAN_SUCCESS on success, non-zero on failure.
+ */
+        template<typename T>
+        RollingMedianResult rolling_median(const std::vector<T> data,
+                                           size_t win_length,
+                                           std::vector<T> &result) {
+            if (win_length == 0) {
+                return ROLLING_MEDIAN_WIN_LENGTH;
+            }
+            OrderedStructs::SkipList::HeadNode<T> sl;
+
+            result.clear();
+            std::vector<T> buffer;
+            for (size_t i = 0; i < data.size(); ++i) {
+                sl.insert(data[i]);
+                if (i >= win_length) {
+                    if (win_length % 2 == 1) {
+                        result.push_back(sl.at(win_length / 2));
+                    } else {
+                        /* Even length so average */
+                        sl.at((win_length - 1) / 2, 2, buffer);
+                        assert(buffer.size() == 2);
+                        result.push_back(buffer[0] / 2 + buffer[1] / 2);
+                    }
+                    sl.remove(data[i - win_length]);
+                }
+            }
+            return ROLLING_MEDIAN_SUCCESS;
+        }
+
+/**
+ * Vector based rolling median.
+ * This always uses the lower bound so works correctly for odd sized window lengths.
+ * For even sized window lengths this chooses the lower value rather than averaging two values.
+ * This is useful for, say, strings that can not be averaged.
+ *
+ * The length of the result is data.size() - win_length
+ *
+ * @tparam T Data type.
+ * @param data Data vector.
+ * @param win_length Window length.
+ * @param result Result vector.
+ * @return ROLLING_MEDIAN_SUCCESS on success, non-zero on failure.
+ */
+        template<typename T>
+        RollingMedianResult rolling_median_lower_bound(const std::vector<T> data,
+                                           size_t win_length,
+                                           std::vector<T> &result) {
+            if (win_length == 0) {
+                return ROLLING_MEDIAN_WIN_LENGTH;
+            }
+            OrderedStructs::SkipList::HeadNode<T> sl;
+
+            result.clear();
+            std::vector<T> buffer;
+            for (size_t i = 0; i < data.size(); ++i) {
+                sl.insert(data[i]);
+                if (i >= win_length) {
+                    /* Choose the lower bound rather than averaging. */
+                    result.push_back(sl.at(win_length / 2));
+                    sl.remove(data[i - win_length]);
+                }
+            }
+            return ROLLING_MEDIAN_SUCCESS;
+        }
+
+/**
  * Return an error code.
  */
 #define ROLLING_MEDIAN_ERROR_CHECK                      \
@@ -106,11 +182,14 @@ namespace OrderedStructs {
 
 /**
  * Rolling median where only the odd mid-index is considered.
- * If the win_length is even then (win_length - 1) / 2 value is used.
- * See even_odd_index() for a different treatment of even lengths.
+ * The (win_length - 1) / 2 value is used.
+ * See even_index() for a different treatment of even lengths.
+ *
  * This is valid for all types T.
  * It is up to the caller to ensure that there is enough space in dest for
  * the results, use dest_size() for this.
+ *
+ * The number of valid values in the result is count - win_length
  *
  * @tparam T Type of the value(s).
  * @param src Source array of values.
@@ -125,10 +204,12 @@ namespace OrderedStructs {
         RollingMedianResult odd_index(const T *src, size_t src_stride,
                                       size_t count, size_t win_length,
                                       T *dest, size_t dest_stride) {
+            assert(win_length % 2 == 1);
+            ROLLING_MEDIAN_ERROR_CHECK;
+
             SkipList::HeadNode<T> sl;
             const T *tail = src;
 
-            ROLLING_MEDIAN_ERROR_CHECK;
             for (size_t i = 0; i < count; ++i) {
                 sl.insert(*src);
                 if (i + 1 >= win_length) {
@@ -142,8 +223,52 @@ namespace OrderedStructs {
             return ROLLING_MEDIAN_SUCCESS;
         }
 
-/*
+/**
+ * Rolling median where only the even mid-index is considered.
+ * Then value is: the mean of the two centre values SkipList[(win_length - 1) / 2] / 2 + SkipList[(win_length - 1) / 2] / 2
+ * See odd_index() for a different treatment of odd lengths.
+ *
+ * This is valid only for types T that can be divided by two.
+ * It is up to the caller to ensure that there is enough space in dest for
+ * the results, use dest_size() for this.
+ *
+ * The number of valid values in the result is count - win_length
+ *
+ * @tparam T Type of the value(s).
+ * @param src Source array of values.
+ * @param src_stride Source stride for 2D arrays.
+ * @param count Number of input values.
+ * @param win_length Window length.
+ * @param dest The destination array.
+ * @param dest_stride The destination stride given a 2D array.
+ * @return The result of the Rolling Median operation as a RollingMedianResult enum.
  */
+        template<typename T>
+        RollingMedianResult even_index(const T *src, size_t src_stride,
+                                       size_t count, size_t win_length,
+                                       T *dest, size_t dest_stride) {
+            assert(win_length % 2 == 0);
+            ROLLING_MEDIAN_ERROR_CHECK;
+
+            SkipList::HeadNode<T> sl;
+            std::vector<T> buffer;
+
+            const T *tail = src;
+            for (size_t i = 0; i < count; ++i) {
+                sl.insert(*src);
+                if (i + 1 >= win_length) {
+                    sl.at((win_length - 1) / 2, 2, buffer);
+                    assert(buffer.size() == 2);
+                    *dest = buffer[0] / 2 + buffer[1] / 2;
+                    dest += dest_stride;
+                    sl.remove(*tail);
+                    tail += src_stride;
+                }
+                src += src_stride;
+            }
+            return ROLLING_MEDIAN_SUCCESS;
+        }
+
 /**
  * Rolling median where the mean of adjacent values is used
  * when the window size is even length.
@@ -164,30 +289,17 @@ namespace OrderedStructs {
         RollingMedianResult even_odd_index(const T *src, size_t src_stride,
                                            size_t count, size_t win_length,
                                            T *dest, size_t dest_stride) {
+            RollingMedianResult ret;
             if (win_length % 2 == 1) {
-                return odd_index(src, src_stride,
+                ret = odd_index(src, src_stride,
+                                count, win_length,
+                                dest, dest_stride);
+            } else {
+                ret = even_index(src, src_stride,
                                  count, win_length,
                                  dest, dest_stride);
-            } else {
-                ROLLING_MEDIAN_ERROR_CHECK;
-                SkipList::HeadNode<T> sl;
-                std::vector<T> buffer;
-
-                const T *tail = src;
-                for (size_t i = 0; i < count; ++i) {
-                    sl.insert(*src);
-                    if (i + 1 >= win_length) {
-                        sl.at((win_length - 1) / 2, 2, buffer);
-                        assert(buffer.size() == 2);
-                        *dest = buffer[0] / 2 + buffer[1] / 2;
-                        dest += dest_stride;
-                        sl.remove(*tail);
-                        tail += src_stride;
-                    }
-                    src += src_stride;
-                }
             }
-            return ROLLING_MEDIAN_SUCCESS;
+            return ret;
         }
 
     } // namespace RollingMedian