+++ tests/benchmarks/test_benchmark_SkipList_threaded.py
@@ -1,9 +1,7 @@
 """Some specific PyObject* tests"""
 import logging
-import random
 import sys
 import threading
-import time
 
 import pytest
 
@@ -53,56 +51,65 @@ def insert_has_remove_float_minimal(sl, value, count):
 
 
 def benchmark_insert_has_remove_float_threaded(skip_list, insert_value, num_inserts, num_threads):
-    threads = [
-        threading.Thread(
-            name='insert_has_remove[{:3d}]'.format(i),
-            target=insert_has_remove_float_minimal,
-            args=(skip_list, insert_value, num_inserts)
-        ) for i in range(num_threads)
-    ]
-    for thread in threads:
-        thread.start()
-    # logging.debug('Waiting for worker threads')
-    main_thread = threading.current_thread()
-    for t in threading.enumerate():
-        if t is not main_thread:
-            t.join()
+    if num_threads > 0:
+        threads = [
+            threading.Thread(
+                name='insert_has_remove[{:3d}]'.format(i),
+                target=insert_has_remove_float_minimal,
+                args=(skip_list, insert_value, num_inserts)
+            ) for i in range(num_threads)
+        ]
+        for thread in threads:
+            thread.start()
+        # logging.debug('Waiting for worker threads')
+        main_thread = threading.current_thread()
+        for t in threading.enumerate():
+            if t is not main_thread:
+                t.join()
+    else:
+        insert_has_remove_float_minimal(skip_list, insert_value, num_inserts)
 
 
 @pytest.mark.slow
 @pytest.mark.parametrize(
-    ('length', 'num_threads'),
+    'num_threads',
     (
-            (1000 * 1000, 1,),
-            (1000 * 1000, 2,),
-            (1000 * 1000, 4,),
-            (1000 * 1000, 8,),
-            (1000 * 1000, 16,),
-            (1000 * 1000, 32,),
-            (1000 * 1000, 64,),
-            (1000 * 1000, 128,),
+            0,
+            1,
+            2,
+            4,
+            8,
+            16,
+            32,
+            64,
+            128,
     ),
-    ids=[
-        'Len_1e6_Threads_001',
-        'Len_1e6_Threads_002',
-        'Len_1e6_Threads_004',
-        'Len_1e6_Threads_008',
-        'Len_1e6_Threads_016',
-        'Len_1e6_Threads_032',
-        'Len_1e6_Threads_064',
-        'Len_1e6_Threads_128',
-    ]
+    # ids=[
+    #     'Len_1e6_Threads_000',
+    #     'Len_1e6_Threads_001',
+    #     'Len_1e6_Threads_002',
+    #     'Len_1e6_Threads_004',
+    #     'Len_1e6_Threads_008',
+    #     'Len_1e6_Threads_016',
+    #     'Len_1e6_Threads_032',
+    #     'Len_1e6_Threads_064',
+    #     'Len_1e6_Threads_128',
+    # ]
 )
-def test_ihr_float_threaded(benchmark, length, num_threads):
-    """Tests multi-threaded insert()/has()/remove() of floats converted to
+def test_ihr_float_threaded(benchmark, num_threads):
+    """Tests multithreaded insert()/has()/remove() of floats converted to
     native C++ objects."""
     # logging.debug('Starting: {}()'.format(sys._getframe().f_code.co_name))
     sl = orderedstructs.SkipList(float)
+    length = 1000 * 1000
     # Load SkipList with length values
     for i in range(length):
         sl.insert(float(i))
     insert_value = float(length / 2.0)
     # Scale the number of repeat inserts to the number of threads.
-    num_inserts = 1024 * 8 // num_threads
+    if num_threads == 0:
+        num_inserts = 1024 * 8
+    else:
+        num_inserts = 1024 * 8 // num_threads
     benchmark(benchmark_insert_has_remove_float_threaded, sl, insert_value, num_inserts, num_threads)
     # logging.debug('Ending: {}()'.format(sys._getframe().f_code.co_name))